refactor(notification): vertical slice architecture for admin controllers
Notification predated the platform's vertical-slice convention: its two
controllers (SmackAdminController, BackfillController) held all logic
inline under a Controllers folder. This moves the service onto the same
Application/{Feature}/{Queries|Commands}/{Name} layout the API uses:

- Six Smack slices: PreviewSmack, GetSmackPhrases, GetSmackRatings
  queries; CreateSmackPhrase, UpdateSmackPhrase, RateSmackPreview
  commands. Shared upsert validation + DTO mapping extracted to
  SmackPhraseUpsertValidation.
- Three Backfill command slices: RequestUsersBackfill,
  RequestPickemGroupsBackfill, RequestPickemGroupMatchupsBackfill,
  each returning Result<Guid> with the correlationId.
- Controllers rewritten as thin shells inside their feature folders:
  [FromServices] handler injection + ToActionResult(). No MediatR by
  design (it is slated for removal platform-wide). Controllers folder
  deleted.
- Handlers return Result<T>; UpdateSmackPhrase maps xmin concurrency
  misses to ResultStatus.Conflict so the 409 contract survives, and
  Backfill preserves its 202 { correlationId } body via the
  ToActionResult onSuccess overload.

Wire-level behavior is unchanged: same routes, same ApiKeyAuth, same
status codes, including the bare 200 on POST ratings, enum IsDefined
loophole guards, trim/length caps, and the 23505 upsert race retry.
One nuance: validation 400s now return the standard { errors } failure
envelope instead of a bare string, matching every other service.

Tests: the 16 controller tests migrated to per-handler files asserting
Result<T> status/value, with the shared fixture (fixed clock, real
SmackPhraseCatalog for preview fidelity, seed helpers) preserved in
SmackTestBase. GetSmackPhrases, previously uncovered, gains a test.
Notification suite: 121 passed.
